Fix:Удалить mvnw,HELP.md,исправить оформление, преобразовать метод валидации

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -1,7 +1,5 @@
 package ru.yandex.practicum.filmorate.model;
 
-
-
 import lombok.Builder;
 import lombok.Data;
 
@@ -11,7 +9,7 @@ import java.time.LocalDate;
 public class Film {
 
 
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private LocalDate releaseDate;
